Markers PoC. Code refactoring.
